* commit 'e5ea6bb8c3261083d6c21ab5f6576f3f11b0b95b': (759 commits)
  chore: delete test file
  chore: test file
  fix: 补全一下，没修改全
  REPORT-63368 修复树组件展开后左边线对不齐问题
  auto upgrade version to 2.0.20220322093515
  无JIRA 补充类型
  auto upgrade version to 2.0.20220321104525
  feature: 支持align
  auto upgrade version to 2.0.20220321101434
  bugfix
  bi.table支持grid布局
  无JIRA任务 chore: 调整npmignore
  auto upgrade version to 2.0.20220317180458
  无jira任务 tab showIndex function v3
  无jira任务 tab showIndex function v2
  无jira任务 tab showIndex function v1
  无jira任务 tab showIndex function
  auto upgrade version to 2.0.20220316144457
  auto upgrade version to 2.0.20220316142521
  无JIRA任务 编号修改
  ...

+++ src/base/combination/tab.js
@@ -15,14 +15,15 @@ BI.Tab = BI.inherit(BI.Widget, {
             tab: false,
             cardCreator: function (v) {
                 return BI.createWidget();
-            }
+            },
+            keepAlives: []
         });
     },
 
     render: function () {
         var self = this, o = this.options;
         if (BI.isObject(o.tab)) {
-            this.tab = BI.createWidget(this.options.tab, {type: "bi.button_group"});
+            this.tab = BI.createWidget(this.options.tab, { type: "bi.button_group" });
             this.tab.on(BI.Controller.EVENT_CHANGE, function (type, value, obj) {
                 self.fireEvent(BI.Controller.EVENT_CHANGE, arguments);
             });
@@ -44,6 +45,7 @@ BI.Tab = BI.inherit(BI.Widget, {
             cardCreator: function (v) {
                 var card = o.cardCreator.apply(self, arguments);
                 self.cardMap[v] = card;
+
                 return card;
             },
             afterCardShow: function (v) {
@@ -60,7 +62,7 @@ BI.Tab = BI.inherit(BI.Widget, {
         var self = this, o = this.options;
         if (o.single === true) {
             BI.each(this.cardMap, function (name, card) {
-                if (name !== (currCardName + "")) {
+                if (name !== (currCardName + "") && self._keepAlive(name) !== true) {
                     self.layout.deleteCardByName(name);
                     delete self.cardMap[name];
                 }
@@ -76,10 +78,23 @@ BI.Tab = BI.inherit(BI.Widget, {
         }
     },
 
-    created: function () {
+    _keepAlive: function (v) {
         var o = this.options;
+
+        return BI.isFunction(o.keepAlives) ? o.keepAlives(v) : BI.contains(o.keepAlives, v);
+    },
+
+    created: function () {
+        var self = this, o = this.options;
         if (o.showIndex !== false) {
-            this.setSelect(o.showIndex);
+            if (BI.isFunction(o.showIndex)) {
+                var v = this.__watch(o.showIndex, function (context, newValue) {
+                    self.setSelect(newValue);
+                });
+                this.setSelect(v);
+            } else {
+                this.setSelect(o.showIndex);
+            }
         }
     },
 
@@ -99,6 +114,7 @@ BI.Tab = BI.inherit(BI.Widget, {
             if (BI.isEqual(name, (cardname + ""))) {
                 self.layout.deleteCardByName(name);
                 delete self.cardMap[name];
+
                 return true;
             }
         });
@@ -118,7 +134,8 @@ BI.Tab = BI.inherit(BI.Widget, {
 
     getTab: function (v) {
         this._assertCard(v);
-        return this.layout.getCardByName(v);
+
+return this.layout.getCardByName(v);
     },
 
     setValue: function (v) {
